requirements.txt
fastapi>=0.115.2
uvicorn[standard]>=0.18.3
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.18 